Delete files

+++ controllers/admin.js
@@ -1,11 +1,13 @@
-const { validationResult } = require("express-validator");
+const fileHelper = require('../util/file');
 
-const Product = require("../models/product");
+const { validationResult } = require('express-validator');
+
+const Product = require('../models/product');
 
 exports.getAddProduct = (req, res, next) => {
-  res.render("admin/edit-product", {
-    pageTitle: "Add Product",
-    path: "/admin/add-product",
+  res.render('admin/edit-product', {
+    pageTitle: 'Add Product',
+    path: '/admin/add-product',
     editing: false,
     hasError: false,
     errorMessage: null,
@@ -20,9 +22,9 @@ exports.postAddProduct = (req, res, next) => {
   const price = req.body.price;
 
   if (!image) {
-    return res.status(422).render("admin/edit-product", {
-      pageTitle: "Add Product",
-      path: "/admin/add-product",
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
       editing: false,
       hasError: true,
       product: {
@@ -39,9 +41,9 @@ exports.postAddProduct = (req, res, next) => {
 
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    return res.status(422).render("admin/edit-product", {
-      pageTitle: "Add Product",
-      path: "/admin/add-product",
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
       editing: false,
       hasError: true,
       product: {
@@ -65,8 +67,8 @@ exports.postAddProduct = (req, res, next) => {
     .save()
     .then(result => {
       //console.log(result);
-      console.log("Created a product");
-      res.redirect("/admin/products");
+      console.log('Created a product');
+      res.redirect('/admin/products');
     })
     .catch(err => {
       const error = new Error(err);
@@ -78,17 +80,17 @@ exports.postAddProduct = (req, res, next) => {
 exports.getEditProduct = (req, res, next) => {
   const editMode = req.query.edit;
   if (!editMode) {
-    res.redirect("/");
+    res.redirect('/');
   }
   const prodId = req.params.productId;
   Product.findById(prodId)
     .then(product => {
       if (!product) {
-        return res.redirect("/");
+        return res.redirect('/');
       }
-      res.render("admin/edit-product", {
-        pageTitle: "Edit Product",
-        path: "/admin/edit-product",
+      res.render('admin/edit-product', {
+        pageTitle: 'Edit Product',
+        path: '/admin/edit-product',
         editing: editMode,
         product: product,
         hasError: false,
@@ -112,9 +114,9 @@ exports.postEditProduct = (req, res, next) => {
 
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    return res.status(422).render("admin/edit-product", {
-      pageTitle: "Edit Product",
-      path: "/admin/edit-product",
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Edit Product',
+      path: '/admin/edit-product',
       editing: true,
       hasError: true,
       product: {
@@ -131,17 +133,18 @@ exports.postEditProduct = (req, res, next) => {
   Product.findById(prodId)
     .then(product => {
       if (product.userId.toString() !== req.user._id.toString()) {
-        return res.redirect("/");
+        return res.redirect('/');
       }
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = updatedImageUrl;
       }
       return product.save().then(result => {
-        console.log("UPDATED PRODUCT!");
-        res.redirect("/admin/products");
+        console.log('UPDATED PRODUCT!');
+        res.redirect('/admin/products');
       });
     })
     .catch(err => {
@@ -157,10 +160,10 @@ exports.getProducts = (req, res, next) => {
     // .populate("userId", "name")
     .then(products => {
       console.log(products);
-      res.render("admin/products", {
+      res.render('admin/products', {
         prods: products,
-        pageTitle: "Admin Products",
-        path: "/admin/products"
+        pageTitle: 'Admin Products',
+        path: '/admin/products'
       });
     })
     .catch(err => {
@@ -172,10 +175,17 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found.'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
-      console.log("DESTROYED PRODUCT");
-      res.redirect("/admin/products");
+      console.log('DESTROYED PRODUCT');
+      res.redirect('/admin/products');
     })
     .catch(err => {
       const error = new Error(err);
